+++ tools/IceRpc.BuildTelemetry.Reporter/Program.cs
@@ -4,8 +4,9 @@ using IceRpc;
 using IceRpc.BuildTelemetry.Reporter;
 using System.Net.Security;
 using System.Security.Cryptography.X509Certificates;
+using System.Reflection;
 
-const var timeout = TimeSpan.FromSeconds(3); // The timeout for the complete telemetry upload process.
+var timeout = TimeSpan.FromSeconds(3); // The timeout for the complete telemetry upload process.
 
 // TODO: Update the URI to the build telemetry server once it is deployed.
 const string uri = "icerpc://localhost"; // The URI of the server.
@@ -44,11 +45,28 @@ try
         .Skip(1)
         .FirstOrDefault() ?? "unknown";
 
+    // Determine the IceRPC version using the assembly version. The assembly version is kept in sync with the
+    // IceRPC version.
+    var assembly = Assembly.GetAssembly(typeof(SliceTelemetry));
+    string version = assembly!.GetName().Version!.ToString();
+
+    // Parse the compilation hash
+    string compilationHash = args
+        .SkipWhile(arg => arg != "--hash")
+        .Skip(1)
+        .FirstOrDefault() ?? "unknown";
+
+    // Parse the contains-slice1 argument
+    bool containsSlice1 = bool.TryParse(args
+        .SkipWhile(arg => arg != "--contains-slice1")
+        .Skip(1)
+        .FirstOrDefault(), out var result) && result;
+
     // Create the appropriate telemetry object based on the IDL
     BuildTelemetry buildTelemetry = idl switch
     {
-        "Slice" => new BuildTelemetry.Slice(new SliceTelemetry(args)),
-        _ => new BuildTelemetry.Protobuf(new ProtobufTelemetry(args))
+        "Slice" => new BuildTelemetry.Slice(new SliceTelemetry(version, compilationHash, containsSlice1)),
+        _ => new BuildTelemetry.Protobuf(new ProtobufTelemetry(version, compilationHash))
     };
 
     // Upload the telemetry to the server.
